requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.19.0
pillow==10.0.0
numpy==1.26.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
opencv-python==4.8.0.76 